+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## <next>
 * Fix Tab issue in editing mode of `Datagrid`
+* Add align to input components of `Datagrid` if prop `align` is set
 
 ## 4.4.0
 * Add `FloatingSelect` and `FloatingSelectCreatable` components